refactor: overall code reorganization
