+++ benches/expand_from_coeff.rs
@@ -1,5 +1,10 @@
 use divan::{black_box, AllocProfiler, Bencher};
-use whir::algebra::{fields::Field64, ntt::NttEngine, random_vector};
+use whir::algebra::{
+    buffer::{ActiveBuffer, BufferOps},
+    fields::Field64,
+    ntt::{Messages, NttEngine, ReedSolomon},
+    random_vector,
+};
 
 #[global_allocator]
 static ALLOC: AllocProfiler = AllocProfiler::system();
@@ -28,19 +33,17 @@ fn interleaved_rs_encode(bencher: Bencher, case: &(usize, usize, usize)) {
             let message_length = 1 << (exp - coset_sz);
             let num_messages = 1 << coset_sz;
             let mut rng = ark_std::rand::thread_rng();
-            let coeffs: Vec<Vec<Field64>> = (0..num_messages)
-                .map(|_| random_vector(&mut rng, message_length))
+            let coeffs: Vec<ActiveBuffer<Field64>> = (0..num_messages)
+                .map(|_| ActiveBuffer::from_vec(random_vector(&mut rng, message_length)))
                 .collect();
             let engine = NttEngine::<Field64>::new_from_fftfield();
             (engine, coeffs, expansion)
         })
         .bench_values(|(engine, coeffs, expansion)| {
-            let coeffs_refs = coeffs.iter().map(|v| v.as_slice()).collect::<Vec<_>>();
-            black_box(engine.interleaved_encode_slices(
-                &coeffs_refs,
-                &[],
-                coeffs[0].len() * expansion,
-            ))
+            let coeffs_refs = coeffs.iter().collect::<Vec<_>>();
+            let messages = Messages::new(&coeffs_refs, coeffs[0].len(), 1);
+            let masks = ActiveBuffer::from_slice(&[]);
+            black_box(engine.interleaved_encode(messages, &masks, coeffs[0].len() * expansion))
         });
 }
