+++ src/bin/main.rs
@@ -6,7 +6,6 @@ use ark_std::rand::distributions::{Distribution, Standard};
 use clap::Parser;
 use whir::{
     algebra::{
-        buffer::ActiveBuffer,
         embedding::{Basefield, Embedding, Identity},
         fields::{Field128, Field192, Field256, Field64, Field64_2, Field64_3},
         linear_form::{Covector, Evaluate, LinearForm, MultilinearExtension},
@@ -18,7 +17,7 @@ use whir::{
     transcript::{codecs::Empty, Codec, DomainSeparator, ProverState, VerifierState},
 };
 
-use whir::buffer::BufferOps;
+use whir::buffer::{ActiveBuffer, BufferOps};
 
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
